// _universal/onyx-stable-20260919/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  adaptiveDispatcher: typeof adaptiveDispatcher;
  adaptiveJobs: typeof adaptiveJobs;
  adaptivePdf: typeof adaptivePdf;
  adaptiveTestProbes: typeof adaptiveTestProbes;
  adaptiveWatchdog: typeof adaptiveWatchdog;
  artifactMutations: typeof artifactMutations;
  buildTranslationPrompt: typeof buildTranslationPrompt;
  crons: typeof crons;
  exportProject: typeof exportProject;
  forensicProbe: typeof forensicProbe;
  generatePdf: typeof generatePdf;
  history: typeof history;
  http: typeof http;
  identity: typeof identity;
  importJob: typeof importJob;
  jobMutations: typeof jobMutations;
  jobProcessing: typeof jobProcessing;
  languageRules: typeof languageRules;
  liveTest: typeof liveTest;
  liveTestStore: typeof liveTestStore;
  mutations: typeof mutations;
  parsePdf: typeof parsePdf;
  pdfLayout: typeof pdfLayout;
  probes: typeof probes;
  queries: typeof queries;
  renderPdfCore: typeof renderPdfCore;
  resumeServerProject: typeof resumeServerProject;
  sourceData: typeof sourceData;
  translateContent: typeof translateContent;
  translateImage: typeof translateImage;
  translateQueue: typeof translateQueue;
  translationConfig: typeof translationConfig;
  translationContract: typeof translationContract;
  translationContractMutation: typeof translationContractMutation;
  upload: typeof upload;
  zipAssembly: typeof zipAssembly;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;